footbot_tg_bot/requirements.txt
aiogram
python-dotenv
mysql-connector-python
